langchain/Youtube Transcript/requirements.txt
python-dotenv
langchain
langchain-groq
youtube-transcript-api
faiss-cpu
streamlit
langchain_community
sentence-transformers
textwrap
pydantic
requests
transformers